+++ sdk/cosmos/azure-cosmos-spark_3/src/main/scala/com/azure/cosmos/CosmosItemSerializerNoExceptionWrapping.scala
@@ -9,4 +9,9 @@ private[cosmos] abstract class CosmosItemSerializerNoExceptionWrapping extends C
     .CosmosItemSerializerHelper
     .getCosmosItemSerializerAccessor
     .setShouldWrapSerializationExceptions(this, false)
+
+  ImplementationBridgeHelpers
+    .CosmosItemSerializerHelper
+    .getCosmosItemSerializerAccessor
+    .setCanSerialize(this, false)
 }