+++ sdk/keyvault/azure-security-keyvault-secrets/src/main/java/com/azure/security/keyvault/secrets/SecretServiceVersion.java
@@ -37,7 +37,11 @@ public enum SecretServiceVersion implements ServiceVersion {
     /**
      * Service version {@code 7.6}.
      */
-    V7_6("7.6");
+    V7_6("7.6"),
+    /**
+     * Service version {@code 2025-07-01}.
+     */
+    V2025_07_01("2025-07-01");
 
     private final String version;
 
@@ -59,6 +63,6 @@ public enum SecretServiceVersion implements ServiceVersion {
      * @return The latest {@link SecretServiceVersion}.
      */
     public static SecretServiceVersion getLatest() {
-        return V7_6;
+        return V2025_07_01;
     }
 }