+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/implementation/models/SearchPostRequest.java
@@ -9,13 +9,9 @@ import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
 import com.azure.json.JsonWriter;
-import com.azure.search.documents.models.HybridSearch;
 import com.azure.search.documents.models.QueryAnswerType;
 import com.azure.search.documents.models.QueryCaptionType;
 import com.azure.search.documents.models.QueryDebugMode;
-import com.azure.search.documents.models.QueryLanguage;
-import com.azure.search.documents.models.QueryRewritesType;
-import com.azure.search.documents.models.QuerySpellerType;
 import com.azure.search.documents.models.QueryType;
 import com.azure.search.documents.models.ScoringStatistics;
 import com.azure.search.documents.models.SearchMode;
@@ -158,18 +154,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
     private SearchMode searchMode;
 
     /*
-     * A value that specifies the language of the search query.
-     */
-    @Generated
-    private QueryLanguage queryLanguage;
-
-    /*
-     * A value that specifies the type of the speller to use to spell-correct individual search query terms.
-     */
-    @Generated
-    private QuerySpellerType querySpeller;
-
-    /*
      * The comma-separated list of fields to retrieve. If unspecified, all fields marked as retrievable in the schema
      * are included.
      */
@@ -233,18 +217,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
     private QueryCaptionType captions;
 
     /*
-     * A value that specifies whether query rewrites should be generated to augment the search query.
-     */
-    @Generated
-    private QueryRewritesType queryRewrites;
-
-    /*
-     * The comma-separated list of field names used for semantic ranking.
-     */
-    @Generated
-    private List<String> semanticFields;
-
-    /*
      * The query parameters for vector and hybrid search queries.
      */
     @Generated
@@ -256,12 +228,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
      */
     @Generated
     private VectorFilterMode vectorFilterMode;
-
-    /*
-     * The query parameters to configure hybrid search behaviors.
-     */
-    @Generated
-    private HybridSearch hybridSearch;
 
     /**
      * Creates an instance of SearchPostRequest class.
@@ -699,52 +665,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
     }
 
     /**
-     * Get the queryLanguage property: A value that specifies the language of the search query.
-     *
-     * @return the queryLanguage value.
-     */
-    @Generated
-    public QueryLanguage getQueryLanguage() {
-        return this.queryLanguage;
-    }
-
-    /**
-     * Set the queryLanguage property: A value that specifies the language of the search query.
-     *
-     * @param queryLanguage the queryLanguage value to set.
-     * @return the SearchPostRequest object itself.
-     */
-    @Generated
-    public SearchPostRequest setQueryLanguage(QueryLanguage queryLanguage) {
-        this.queryLanguage = queryLanguage;
-        return this;
-    }
-
-    /**
-     * Get the querySpeller property: A value that specifies the type of the speller to use to spell-correct individual
-     * search query terms.
-     *
-     * @return the querySpeller value.
-     */
-    @Generated
-    public QuerySpellerType getQuerySpeller() {
-        return this.querySpeller;
-    }
-
-    /**
-     * Set the querySpeller property: A value that specifies the type of the speller to use to spell-correct individual
-     * search query terms.
-     *
-     * @param querySpeller the querySpeller value to set.
-     * @return the SearchPostRequest object itself.
-     */
-    @Generated
-    public SearchPostRequest setQuerySpeller(QuerySpellerType querySpeller) {
-        this.querySpeller = querySpeller;
-        return this;
-    }
-
-    /**
      * Get the select property: The comma-separated list of fields to retrieve. If unspecified, all fields marked as
      * retrievable in the schema are included.
      *
@@ -967,52 +887,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
     }
 
     /**
-     * Get the queryRewrites property: A value that specifies whether query rewrites should be generated to augment the
-     * search query.
-     *
-     * @return the queryRewrites value.
-     */
-    @Generated
-    public QueryRewritesType getQueryRewrites() {
-        return this.queryRewrites;
-    }
-
-    /**
-     * Set the queryRewrites property: A value that specifies whether query rewrites should be generated to augment the
-     * search query.
-     *
-     * @param queryRewrites the queryRewrites value to set.
-     * @return the SearchPostRequest object itself.
-     */
-    @Generated
-    public SearchPostRequest setQueryRewrites(QueryRewritesType queryRewrites) {
-        this.queryRewrites = queryRewrites;
-        return this;
-    }
-
-    /**
-     * Get the semanticFields property: The comma-separated list of field names used for semantic ranking.
-     *
-     * @return the semanticFields value.
-     */
-    @Generated
-    public List<String> getSemanticFields() {
-        return this.semanticFields;
-    }
-
-    /**
-     * Set the semanticFields property: The comma-separated list of field names used for semantic ranking.
-     *
-     * @param semanticFields the semanticFields value to set.
-     * @return the SearchPostRequest object itself.
-     */
-    @Generated
-    public SearchPostRequest setSemanticFields(List<String> semanticFields) {
-        this.semanticFields = semanticFields;
-        return this;
-    }
-
-    /**
      * Get the vectorQueries property: The query parameters for vector and hybrid search queries.
      *
      * @return the vectorQueries value.
@@ -1059,28 +933,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
     }
 
     /**
-     * Get the hybridSearch property: The query parameters to configure hybrid search behaviors.
-     *
-     * @return the hybridSearch value.
-     */
-    @Generated
-    public HybridSearch getHybridSearch() {
-        return this.hybridSearch;
-    }
-
-    /**
-     * Set the hybridSearch property: The query parameters to configure hybrid search behaviors.
-     *
-     * @param hybridSearch the hybridSearch value to set.
-     * @return the SearchPostRequest object itself.
-     */
-    @Generated
-    public SearchPostRequest setHybridSearch(HybridSearch hybridSearch) {
-        this.hybridSearch = hybridSearch;
-        return this;
-    }
-
-    /**
      * {@inheritDoc}
      */
     @Generated
@@ -1119,8 +971,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
                     .collect(Collectors.joining(",")));
         }
         jsonWriter.writeStringField("searchMode", this.searchMode == null ? null : this.searchMode.toString());
-        jsonWriter.writeStringField("queryLanguage", this.queryLanguage == null ? null : this.queryLanguage.toString());
-        jsonWriter.writeStringField("speller", this.querySpeller == null ? null : this.querySpeller.toString());
         if (this.select != null) {
             jsonWriter.writeStringField("select",
                 this.select.stream().map(element -> element == null ? "" : element).collect(Collectors.joining(",")));
@@ -1134,17 +984,9 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
         jsonWriter.writeStringField("semanticQuery", this.semanticQuery);
         jsonWriter.writeStringField("answers", this.answers == null ? null : this.answers.toString());
         jsonWriter.writeStringField("captions", this.captions == null ? null : this.captions.toString());
-        jsonWriter.writeStringField("queryRewrites", this.queryRewrites == null ? null : this.queryRewrites.toString());
-        if (this.semanticFields != null) {
-            jsonWriter.writeStringField("semanticFields",
-                this.semanticFields.stream()
-                    .map(element -> element == null ? "" : element)
-                    .collect(Collectors.joining(",")));
-        }
         jsonWriter.writeArrayField("vectorQueries", this.vectorQueries, (writer, element) -> writer.writeJson(element));
         jsonWriter.writeStringField("vectorFilterMode",
             this.vectorFilterMode == null ? null : this.vectorFilterMode.toString());
-        jsonWriter.writeJsonField("hybridSearch", this.hybridSearch);
         return jsonWriter.writeEndObject();
     }
 
@@ -1217,10 +1059,6 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
                     deserializedSearchPostRequest.searchFields = searchFields;
                 } else if ("searchMode".equals(fieldName)) {
                     deserializedSearchPostRequest.searchMode = SearchMode.fromString(reader.getString());
-                } else if ("queryLanguage".equals(fieldName)) {
-                    deserializedSearchPostRequest.queryLanguage = QueryLanguage.fromString(reader.getString());
-                } else if ("speller".equals(fieldName)) {
-                    deserializedSearchPostRequest.querySpeller = QuerySpellerType.fromString(reader.getString());
                 } else if ("select".equals(fieldName)) {
                     List<String> select = reader.getNullable(nonNullReader -> {
                         String selectEncodedAsString = nonNullReader.getString();
@@ -1247,23 +1085,11 @@ public final class SearchPostRequest implements JsonSerializable<SearchPostReque
                     deserializedSearchPostRequest.answers = QueryAnswerType.fromString(reader.getString());
                 } else if ("captions".equals(fieldName)) {
                     deserializedSearchPostRequest.captions = QueryCaptionType.fromString(reader.getString());
-                } else if ("queryRewrites".equals(fieldName)) {
-                    deserializedSearchPostRequest.queryRewrites = QueryRewritesType.fromString(reader.getString());
-                } else if ("semanticFields".equals(fieldName)) {
-                    List<String> semanticFields = reader.getNullable(nonNullReader -> {
-                        String semanticFieldsEncodedAsString = nonNullReader.getString();
-                        return semanticFieldsEncodedAsString.isEmpty()
-                            ? new LinkedList<>()
-                            : new LinkedList<>(Arrays.asList(semanticFieldsEncodedAsString.split(",", -1)));
-                    });
-                    deserializedSearchPostRequest.semanticFields = semanticFields;
                 } else if ("vectorQueries".equals(fieldName)) {
                     List<VectorQuery> vectorQueries = reader.readArray(reader1 -> VectorQuery.fromJson(reader1));
                     deserializedSearchPostRequest.vectorQueries = vectorQueries;
                 } else if ("vectorFilterMode".equals(fieldName)) {
                     deserializedSearchPostRequest.vectorFilterMode = VectorFilterMode.fromString(reader.getString());
-                } else if ("hybridSearch".equals(fieldName)) {
-                    deserializedSearchPostRequest.hybridSearch = HybridSearch.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }