+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/SessionsMessagePump.java
@@ -33,6 +33,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Locale;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
@@ -679,11 +680,11 @@ final class SessionsMessagePump {
         }
 
         private void track(ServiceBusSessionReactorReceiver receiver) {
-            receivers.put(receiver.getSessionId(), receiver);
+            receivers.put(receiver.getSessionId().toLowerCase(Locale.ROOT), receiver);
         }
 
         private void untrack(ServiceBusSessionReactorReceiver receiver) {
-            receivers.remove(receiver.getSessionId(), receiver);
+            receivers.remove(receiver.getSessionId().toLowerCase(Locale.ROOT), receiver);
         }
 
         private void clear() {
@@ -767,7 +768,7 @@ final class SessionsMessagePump {
                 return Mono.error(new UnsupportedOperationException(m));
             }
             final String sessionId = message.getSessionId();
-            final ServiceBusSessionReactorReceiver receiver = receivers.get(sessionId);
+            final ServiceBusSessionReactorReceiver receiver = receivers.get(sessionId.toLowerCase(Locale.ROOT));
             final DeliveryState deliveryState = MessageUtils.getDeliveryState(dispositionStatus, deadLetterReason,
                 deadLetterDescription, propertiesToModify, transactionContext);